refactor: refactored todoDetailModal
